+++ src/main/java/by/itacademy/javaenterprise/borisevich/entity/ExerciseName.java
@@ -1,0 +1,15 @@
+package by.itacademy.javaenterprise.borisevich.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class ExerciseName {
+    private int exerciseNameId;
+    private String exerciseName;
+}
