+++ src/main/java/by/itacademy/javaenterprise/borisevich/entity/TestRdbsConfiguration.java
@@ -1,0 +1,4 @@
+package by.itacademy.javaenterprise.borisevich.entity;
+
+public class TestRdbsConfiguration {
+}
