+++ src/main/java/by/itacademy/javaenterprise/borisevich/exception/DAOException.java
@@ -1,0 +1,21 @@
+package by.itacademy.javaenterprise.borisevich.exception;
+
+public class DAOException extends Exception {
+
+	private static final long serialVersionUID = 1L;
+
+	public DAOException() {
+	}
+
+	public DAOException(String message, Throwable cause) {
+		super(message, cause);
+	}
+
+	public DAOException(String message) {
+		super(message);
+	}
+
+	public DAOException(Throwable cause) {
+		super(cause);
+	}
+}
