--- src/main/java/by/itacademy/javaenterprise/command/Command.java
@@ -1,10 +1,0 @@
-package by.itacademy.javaenterprise.command;
-
-
-import by.itacademy.javaenterprise.exception.DAOException;
-
-public interface Command<T> {
-
-    void execute(T smth) throws DAOException;
-
-}
