--- src/main/java/by/itacademy/javaenterprise/dao/MuscleDAO.java
@@ -1,16 +1,0 @@
-package by.itacademy.javaenterprise.dao;
-
-import by.itacademy.javaenterprise.entity.Muscle;
-import by.itacademy.javaenterprise.exception.DAOException;
-
-import java.util.List;
-
-public interface MuscleDAO<T> {
-    void addMuscle(Muscle muscle) throws DAOException;
-
-    void updateMuscle(Muscle muscle, int muscleId) throws DAOException;
-
-    void deleteMuscle(String muscleName) throws DAOException;
-
-    List<T> getAllMusclesPagination(int limit, int offset) throws DAOException;
-}
