--- src/main/java/by/itacademy/javaenterprise/entity/ExerciseName.java
@@ -1,15 +1,0 @@
-package by.itacademy.javaenterprise.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class ExerciseName {
-    private int exerciseNameId;
-    private String exerciseName;
-}
