--- src/main/java/by/itacademy/javaenterprise/entity/Muscle.java
@@ -1,15 +1,0 @@
-package by.itacademy.javaenterprise.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class Muscle {
-    private int muscleId;
-    private String muscleName;
-}
