--- src/main/java/by/itacademy/javaenterprise/entity/TestRdbsConfiguration.java
@@ -1,4 +1,0 @@
-package by.itacademy.javaenterprise.entity;
-
-public class TestRdbsConfiguration {
-}
