--- src/main/java/by/itacademy/javaenterprise/entity/Training.java
@@ -1,20 +1,0 @@
-package by.itacademy.javaenterprise.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@NoArgsConstructor
-@Builder
-@AllArgsConstructor
-@Data
-public class Training {
-
-    private int id;
-    private LocalDateTime date;
-    private int userID;
-
-}
