+++ src/main/resources/database.properties
@@ -7,3 +7,4 @@ dataSource.prepStmtCacheSize=250
 dataSource.prepStmtCacheSqlLimit=2048
 dataSource.maxPoolSize=20
 dataSource.timeout = 10
+
